baseline/src/requirements.txt
pandas==0.18.1
pyannote-algorithms==0.6.4
